fix(80a): TestClient loopback patch 上移根 conftest + help i18n 子字串守衛健化
全套跑時抓到兩個 LAN-gate 衍生問題：
1. 測試 isolation 403：TestClient loopback 預設 client patch 原在 tests/integration/
   conftest，unit 測試單檔 isolation 跑（CLAUDE.md 開發流程）時未載入 → 打路由的
   unit 測試被閘門 403。上移至根 tests/conftest.py，整套跑與單檔跑行為一致。
2. help 守衛誤觸：feature/80 新增 i18n key settings.server_info.no_lan_ip 含子字串
   "lan_ip"，被 dump 進 window.__i18n 全域 bootstrap，誤觸 38b 的 test_help_page_
   no_lan_ip_reference（blanket "lan_ip" not in html）。排除 window.__i18n 行後再比對，
   保留守衛真意（help 模板不得復活 lan_ip context 變數）、去除無關 key 誤報。

全套 4436 passed, 2 skipped；help/capabilities 等 unit 測試 isolation 亦綠。
